can't update the note, if it contains > ' < twice. id (to update the note) is taken by the note that is sent from RVAdapter to NoteDialogFragment. id is found by the note, so now, it doesn't matter if the id stays the same as earlier and it doesn't confuse the rv

+++ app/src/main/java/com/example/toja/notepad/database/DatabaseHelper.java
@@ -37,7 +37,7 @@ public class DatabaseHelper extends SQLiteOpenHelper {
         return result != -1;
     }
 
-    public boolean updateData(int id, String note) {
+    public boolean updateData(long id, String note) {
         SQLiteDatabase database = this.getWritableDatabase();
         ContentValues contentValues = new ContentValues();
         contentValues.put(Note.COL_ID, id);
